flask/requirements.txt
bidict==0.23.1
blinker==1.8.2
click==8.1.7
dnspython==2.6.1
Flask==3.0.3
Flask-Cors==5.0.0
Flask-SocketIO==5.3.7
h11==0.14.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
pymongo==4.9.1
python-dotenv==1.0.1
python-engineio==4.9.1
python-socketio==5.11.4
simple-websocket==1.0.0
Werkzeug==3.0.4
wsproto==1.2.0
gunicorn==20.1.0